+++ src/test/java/com/fasterxml/jackson/databind/ser/TestCyclicTypes.java
@@ -3,8 +3,6 @@ package com.fasterxml.jackson.databind.ser;
 import java.util.*;
 
 
-import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.fasterxml.jackson.databind.*;
 
 /**
@@ -22,31 +20,21 @@ public class TestCyclicTypes
     /**********************************************************
      */
 
+    static class Bean
+    {
+        Bean _next;
+        final String _name;
 
-  @JsonIdentityInfo(
-      generator = ObjectIdGenerators.PropertyGenerator.class
-      , property = "id"
-      , scope = Bean.class
-  )
-  static class Bean {
-    final int _id;
-    Bean _next;
-    final String _name;
+        public Bean(Bean next, String name) {
+            _next = next;
+            _name = name;
+        }
 
-    public Bean(int id, Bean next, String name) {
-      _id = id;
-      _next = next;
-      _name = name;
+        public Bean getNext() { return _next; }
+        public String getName() { return _name; }
+
+        public void assignNext(Bean n) { _next = n; }
     }
-
-    public int getId() { return _id; }
-
-    public Bean getNext() { return _next; }
-
-    public String getName() { return _name; }
-
-    public void assignNext(Bean n) { _next = n; }
-  }
 
     /*
     /**********************************************************
@@ -54,30 +42,37 @@ public class TestCyclicTypes
     /**********************************************************
      */
 
-  public void testLinked() throws Exception {
-    Bean sameChild = new Bean(3, null, "sameChild");
-    Bean first = new Bean(1, sameChild, "first");
-    Bean second = new Bean(2, sameChild, "second");
-    sameChild.assignNext(first);
+    public void testLinked() throws Exception
+    {
+        Bean last = new Bean(null, "last");
+        Bean first = new Bean(last, "first");
+        Map<String,Object> map = writeAndMap(new ObjectMapper(), first);
 
-    final ObjectMapper mapper = new ObjectMapper();
-    mapper.enable(SerializationFeature.HANDLE_CIRCULAR_REFERENCE_INDIVIDUALLY_FOR_COLLECTIONS);
-    final String result = mapper.writeValueAsString(Arrays.asList(first, second));
+        assertEquals(2, map.size());
+        assertEquals("first", map.get("name"));
 
-    final StringBuilder expected = new StringBuilder();
-    expected.append("[");
-    expected.append("{\"id\":1,\"name\":\"first\",\"next\":");
-    expected.append(
-        "{\"id\":3,\"name\":\"sameChild\",\"next\":1}}"); // 1 has been visited this iteration => next is a reference
-    expected.append(",{\"id\":2,\"name\":\"second\",\"next\":");
-    expected.append(
-        "{\"id\":3,\"name\":\"sameChild\",\"next\":"); // 1 has noy been visited this iteration => next is fully serialized
-    expected
-        .append("{\"id\":1,\"name\":\"first\",\"next\":3}"); // 3 has been visited this iteration => next is a reference
-    expected.append("}");
-    expected.append("}");
-    expected.append("]");
+        @SuppressWarnings("unchecked")
+        Map<String,Object> map2 = (Map<String,Object>) map.get("next");
+        assertNotNull(map2);
+        assertEquals(2, map2.size());
+        assertEquals("last", map2.get("name"));
+        assertNull(map2.get("next"));
+    }
 
-    assertEquals(result, expected.toString());
-  }
+    /**
+     * Test for verifying that [JACKSON-158] works as expected
+     */
+    public void testSelfReference() throws Exception
+    {
+        Bean selfRef = new Bean(null, "self-refs");
+        Bean first = new Bean(selfRef, "first");
+        selfRef.assignNext(selfRef);
+        ObjectMapper m = new ObjectMapper();
+        Bean[] wrapper = new Bean[] { first };
+        try {
+            writeAndMap(m, wrapper);
+        } catch (JsonMappingException e) {
+            verifyException(e, "Direct self-reference leading to cycle");
+        }
+    }
 }
